Initialization of subject and creating answer
Implemented a simple version of initializing a subject either by using one of the test subjects, or by randomizing a new subject. When randomizing, only tones in the chosen scale will be used.

Also implemented a function that creates the answer by transposing to the dominant and then checking if it should be set down one octave.

Created classes for a Note and a Scale, and function for checking if the pitch of a note is in the scale.