Object oriented creation of the base fugue
I made the program more object oriented and made a Fugue object that when initialized creates the subject and all the parts that is only transposed versions of the subject, according to the guide we found on how to create a short 12 bar fugue. So using the guide we found, this means that:

Bar 1, 9 and 10 are completely done.
Bar 2 and 6 are only missing a countersubject for first voice (same in both but transposed)
Bar 5 is missing something for second voice.
Bar 3, 4, 7, 8, 11 and 12 are not touched yet.