Evolutionary Algorithm
Made an evolutionary algorithm that later can be combined with the rest of the program. Right now it only generates a melody so close to the tone c1 as possible. Next step is to decide what kind of melodies we want it to generate, and how to create a fitness function that gives us what we want.