Solved the transposing rests problem
Fixed so that it is is now possible to transpose tracks containing rests when using the function transpose_from_halfnote. Also fixed some small stuff like spelling errors, changed __len__ to len, changed an index and added copy instead of pointing at same object. Oh, and learned that it is called rests and not pauses so changed that to.