Moved add_tracks and translated cross_over
I moved the add_tracks function from newFugue to track_functions to make it accessible from EG too. Also continued with translating EG to mingus and am done with cross-over now. I come accross a lot of problems with mingus not having the functionality I want, or maybe I just don't find it, anyway it causes everything to take a lot of time to translate...