EG running smoothly! Two different test fitness functions.
Bug in cross-over finally found, so now everything runs smoothly! Had forgotten to increase an index in a while-loop, which caused strange behaviour that sometimes worked without error (but not correct cross-over), and sometimes gave running error.

Made so that you can choose which fitness function to use for the EG. Right now there is only two possible.