Small suggestion on part of a fitnessfunction
I tried to write  (a start of) a fitness function. It works, but should be one of several 'tests'. And I don't know any common practice in how to write a fitness function, how to distribute 'points' and so on, so there are probably better ways to do it. Just wanted to share what I played around with today