Finishing first draft of contrapunctal motion
Can now measure contrapunctal motion for two tracks. Gives a dictionary with the different motion types as keys and the values set to the percentage of how much each type is used. The measure isn't perfect in the switches between different types, so that could be improved.

Also moved the function interval_at_beat to track_functions because that felt more suitable. Should create a function in track_tests that uses that one to measure intervals at the four beats in each bar though!