Fitness function for harmonies
Been working on a fitness function for creating a harmony to an existing input melody. It creates very interesting results. One example can be seen in the new folder, made for saving fugues from different stages in the progress of the fugue generator.

contrapunctal_motion and its help function still got some problems, so it's not part of the fitness function yet.