Simplifying note names
Fixed so that Gbbb will be named E and so on. Otherwise two fitness test functions will give incorrect answers. I think this function might cause problems when rendering a pdf, not sure if it is this or some other change that creates the problem. Also found a small bug and changed the number of generations to a slightly higher number. 1000 would be great, but to keep down the run time I chose 500 for now.