Changing the ending
Created a new version of the ending. Now the second voice plays a cadence of base notes, while the first voice is generated with its own fitness function, making it care about both binding together the bar before and after (which is just the root node), and care about harmonizing the second voice well. No finetuning of parameters is done except changing the weights for different note durations to prefer quarter and half notes more and short notes less.